refactor(db): reorganise schema, split project_file
Group the flat schema files into auth/, ai/, billing/ and projects/, each exported as a subpath so consumers import only what they need.

Split the four large jsonb columns (scene, spec, content, history) out of project_file into a 1:1 project_file_content. project_file was almost always TOASTed, so the two hottest queries in the app, the dashboard tree and the workspace file list, shared pages and cache with data neither of them reads. Reads left-join and writes upsert, so a file whose content row is missing degrades to empty rather than vanishing from a list.

Drop the waitlist table (preparing for launch).
