+++ packages/db/src/schema/ai/index.ts
@@ -1,0 +1,1 @@
+export * from "./user-ai-provider";
