+++ packages/db/src/schema/auth/better-auth.ts
@@ -73,17 +73,9 @@ export const verification = pgTable(
   (table) => [index("verification_identifier_idx").on(table.identifier)],
 );
 
-/**
- * Better Auth's rate-limit counters.
- *
- * Exists so `rateLimit.storage` can be `"database"`. The default is in-process
- * memory, which on Cloud Run means the limit is per instance and is wiped by every
- * scale-to-zero -- so the brute-force protection on `/sign-in` lapses whenever the
- * service goes idle, which for a low-traffic deployment is most of the time.
- *
- * Shape is dictated by better-auth (`getAuthTables`): `key` unique, `count`, and
- * `lastRequest` as epoch millis in a bigint. Written and read only by better-auth.
- */
+// Better Auth's rate-limit counter Table.
+// TODO: we can delete it, if we ever move to a vps for api
+// Ref: https://better-auth.com/docs/concepts/rate-limit#schema
 export const rateLimit = pgTable("rate_limit", {
   id: text("id").primaryKey(),
   key: text("key").notNull().unique(),
