+++ packages/db/src/schema/index.ts
@@ -1,11 +1,4 @@
 export * from "./auth";
-export * from "./creation-usage";
-export * from "./github-import-job";
-export * from "./plan";
-export * from "./project";
-export * from "./project-file";
-export * from "./subscription";
-export * from "./usage-ledger";
-export * from "./user-ai-provider";
-export * from "./waitlist";
-export * from "./webhook-event";
+export * from "./ai";
+export * from "./billing";
+export * from "./projects";
