+++ packages/db/src/schema/projects/index.ts
@@ -1,0 +1,4 @@
+export * from "./project";
+export * from "./project-file";
+export * from "./project-file-content";
+export * from "./github-import-job";
