+++ packages/db/src/schema/projects/project-file.ts
@@ -1,7 +1,8 @@
 import { relations, sql } from "drizzle-orm";
-import { index, jsonb, pgTable, text, timestamp, check } from "drizzle-orm/pg-core";
+import { index, pgTable, text, timestamp, check } from "drizzle-orm/pg-core";
 
 import { project } from "./project";
+import { projectFileContent } from "./project-file-content";
 
 export const projectFileTypes = ["diagram", "doc"] as const;
 
@@ -16,12 +17,10 @@ export const projectFile = pgTable(
       .references(() => project.id, { onDelete: "cascade" }),
     type: text("type", { enum: projectFileTypes }).notNull(),
     name: text("name").notNull(),
-    scene: jsonb("scene"),
-    spec: jsonb("spec"),
-    content: jsonb("content"),
-    history: jsonb("history")
-      .$default(() => [])
-      .notNull(),
+    // `scene`, `spec`, `content` and `history` live in `project_file_content`,
+    // keyed 1:1 by file id. See that file for why. This table is deliberately
+    // narrow: it is what the dashboard tree and the file list read, and neither
+    // of them wants any of the large columns.
     createdAt: timestamp("created_at").defaultNow().notNull(),
     updatedAt: timestamp("updated_at")
       .defaultNow()
@@ -30,7 +29,9 @@ export const projectFile = pgTable(
   },
   (table) => [
     index("project_file_projectId_idx").on(table.projectId),
-    index("project_file_type_idx").on(table.type),
+    // Deliberately no index on `type`. It is only ever selected, never filtered,
+    // and with two distinct values a btree would be ignored even if it were --
+    // the check constraint below is what actually guards the column.
     check("project_file_type_check", sql`${table.type} IN ('diagram', 'doc')`),
   ],
 );
@@ -39,5 +40,9 @@ export const projectFileRelations = relations(projectFile, ({ one }) => ({
   project: one(project, {
     fields: [projectFile.projectId],
     references: [project.id],
+  }),
+  content: one(projectFileContent, {
+    fields: [projectFile.id],
+    references: [projectFileContent.fileId],
   }),
 }));
