--- packages/db/src/schema/waitlist.ts
@@ -1,18 +1,0 @@
-import { pgTable, text, timestamp, uniqueIndex } from "drizzle-orm/pg-core";
-import { user } from "./auth";
-
-export const waitlist = pgTable(
-  "waitlist",
-  {
-    id: text("id")
-      .primaryKey()
-      .$defaultFn(() => crypto.randomUUID()),
-    userId: text("user_id").references(() => user.id, { onDelete: "cascade" }),
-    email: text("email").notNull(),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-  },
-  (table) => [
-    uniqueIndex("waitlist_email_idx").on(table.email),
-    uniqueIndex("waitlist_user_id_idx").on(table.userId),
-  ],
-);
